x86: Cast function pointer to unsigned long in install_signal_bh()
Fixes the following compilation warning:

    CC       arch/x86/signal.o
  arch/x86/signal.c: In function ‘install_signal_bh’:
  arch/x86/signal.c:78: warning: assignment makes integer from pointer without a cast

+++ arch/x86/signal.c
@@ -75,7 +75,7 @@ int install_signal_bh(void *ctx, signal_bh_fn bh)
 
 	/* push bottom-half handler address on stack */
 	stack--;
-	*stack = bh;
+	*stack = (unsigned long) bh;
 
 	uc->uc_mcontext.gregs[REG_SP] -= 2 * sizeof(unsigned long);
